+++ app/audio/analyzers/hpcp_extended.py
@@ -48,4 +48,4 @@ class HpCPExtendedAnalyzer(BaseAnalyzer):
         return {"hpcp_entropy": entropy, "hpcp_crest": crest}
 
     def _extract(self, ctx: AnalysisContext) -> dict[str, Any]:
-        return self.analyze(ctx.samples, ctx.sr)
+        return self.analyze(ctx.samples.astype(np.float32, copy=False), ctx.sr)
